# web-server/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Meeting Transcription API",
    "version": "1.0.0",
    "description": "A comprehensive API for managing meetings and their transcriptions with real-time audio processing.",
    "contact": {
      "name": "Meeting Transcription Team",
      "email": "support@transcription.local"
    },
    "license": {
      "name": "MIT License"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/transcription/job/{job_id}": {
      "get": {
        "tags": ["transcription"],
        "summary": "Get transcription job status",
        "description": "Get the status of a specific transcription job",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the transcription job"
          }
        ],
        "responses": {
          "200": {
            "description": "Job status information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "job_id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": ["queued", "processing", "completed", "failed"]
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "completed_at": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    },
                    "error_message": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Transcription job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          }
        }
      }
    },
    "/meetings": {
      "get": {
        "tags": ["meetings"],
        "summary": "Get all meetings",
        "description": "Retrieve all meetings for the current user",
        "responses": {
          "200": {
            "description": "List of meetings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Meeting"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["meetings"],
        "summary": "Create a new meeting",
        "description": "Create a new meeting with title, description, and keywords",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeetingCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Meeting created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meeting"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          }
        }
      }
    },
    "/meetings/{meeting_id}": {
      "get": {
        "tags": ["meetings"],
        "summary": "Get a specific meeting",
        "description": "Retrieve a meeting by its ID",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the meeting"
          }
        ],
        "responses": {
          "200": {
            "description": "Meeting details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meeting"
                }
              }
            }
          },
          "404": {
            "description": "Meeting not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["meetings"],
        "summary": "Update a meeting",
        "description": "Update meeting details",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the meeting"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeetingUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Meeting updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meeting"
                }
              }
            }
          },
          "404": {
            "description": "Meeting not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["meetings"],
        "summary": "Delete a meeting",
        "description": "Delete a meeting by its ID",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the meeting"
          }
        ],
        "responses": {
          "200": {
            "description": "Meeting deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Meeting deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Meeting not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          }
        }
      }
    },
    "/meetings/{meeting_id}/keywords": {
      "put": {
        "tags": ["meetings"],
        "summary": "Update meeting keywords",
        "description": "Update the keywords for a specific meeting",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the meeting"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeywordsUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Keywords updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meeting"
                }
              }
            }
          },
          "404": {
            "description": "Meeting not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          }
        }
      }
    },
    "/webhook/transcription-completed": {
      "post": {
        "tags": ["webhooks"],
        "summary": "Transcription webhook",
        "description": "Receive transcription results from the transcription service",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranscriptionWebhookResult"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook processed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Webhook processed"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid webhook data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Meeting": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the meeting"
          },
          "title": {
            "type": "string",
            "description": "Meeting title"
          },
          "description": {
            "type": "string",
            "description": "Meeting description",
            "nullable": true
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Keywords associated with the meeting"
          },
          "status": {
            "type": "string",
            "enum": ["created", "uploading", "transcribing", "completed", "failed"],
            "description": "Current status of the meeting"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Meeting creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Meeting last update timestamp"
          },
          "fullTranscription": {
            "type": "string",
            "description": "Full transcription text",
            "nullable": true
          },
          "metadata": {
            "type": "object",
            "description": "Additional meeting metadata",
            "nullable": true,
            "properties": {
              "language": {
                "type": "string"
              },
              "participants": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "totalDuration": {
                "type": "number"
              },
              "processingStarted": {
                "type": "string",
                "format": "date-time"
              },
              "processingCompleted": {
                "type": "string",
                "format": "date-time"
              },
              "totalProcessingTime": {
                "type": "number"
              }
            }
          }
        },
        "required": ["id", "title", "status", "created_at", "updated_at"]
      },
      "MeetingCreate": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Meeting title"
          },
          "description": {
            "type": "string",
            "description": "Meeting description",
            "nullable": true
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Keywords associated with the meeting",
            "default": []
          }
        },
        "required": ["title"]
      },
      "MeetingUpdate": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Meeting title",
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "Meeting description",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": ["created", "uploading", "transcribing", "completed", "failed"],
            "description": "Meeting status",
            "nullable": true
          }
        }
      },
      "KeywordsUpdate": {
        "type": "object",
        "properties": {
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Updated keywords for the meeting"
          }
        },
        "required": ["keywords"]
      },
...
      "HTTPError": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string",
            "description": "Error message"
          }
        }
      }
    }
  }
}